Rename hats module and triage endpoints
