+++ tests/api/test_triage_api.py
@@ -1,4 +1,5 @@
 import json
+from pathlib import Path
 
 import pytest
 from fastapi.testclient import TestClient
@@ -6,7 +7,9 @@ from fastapi.testclient import TestClient
 from apps.api.main import app
 
 
-def test_post_kpi_and_idempotent(tmp_path, monkeypatch):
+def test_post_kpi_and_idempotent(
+    tmp_path: Path, monkeypatch: pytest.MonkeyPatch
+) -> None:
     ledger = tmp_path / "ledger.jsonl"
     snapshot = tmp_path / "snapshot.json"
     monkeypatch.setenv("HATS_LEDGER_PATH", str(ledger))
@@ -15,7 +18,7 @@ def test_post_kpi_and_idempotent(tmp_path, monkeypatch):
     client = TestClient(app)
     payload = {"wo_id": "1", "hat_id": "h1", "SA": 0.8, "SP": 0.9}
 
-    res = client.post("/hats/kpi", json=payload)
+    res = client.post("/triage/kpi", json=payload)
     assert res.status_code == 200
     data = res.json()
     assert data["hat_id"] == "h1"
@@ -30,18 +33,18 @@ def test_post_kpi_and_idempotent(tmp_path, monkeypatch):
     assert snap  # not empty
 
     # Second post is idempotent
-    res2 = client.post("/hats/kpi", json=payload)
+    res2 = client.post("/triage/kpi", json=payload)
     assert res2.status_code == 200
     assert res2.json() == data
     assert len(ledger.read_text().strip().splitlines()) == 1
 
 
-def test_post_kpi_bad_payload(monkeypatch, tmp_path):
+def test_post_kpi_bad_payload(monkeypatch: pytest.MonkeyPatch, tmp_path: Path) -> None:
     ledger = tmp_path / "ledger.jsonl"
     snapshot = tmp_path / "snapshot.json"
     monkeypatch.setenv("HATS_LEDGER_PATH", str(ledger))
     monkeypatch.setenv("HATS_SNAPSHOT_PATH", str(snapshot))
 
     client = TestClient(app)
-    res = client.post("/hats/kpi", json={"wo_id": "1"})
+    res = client.post("/triage/kpi", json={"wo_id": "1"})
     assert res.status_code == 422
